% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descargar_datos_acciones_eikon.R
\name{descargar_datos_acciones_eikon}
\alias{descargar_datos_acciones_eikon}
\title{Descargar datos de acciones vía Refinitiv Eikon, desde fecha especificada hasta hoy.}
\usage{
descargar_datos_acciones_eikon(
  rics,
  start_date_eikon_df,
  verbose = FALSE,
  raw_output = FALSE
)
}
\arguments{
\item{rics}{lista de rics a descargar}

\item{start_date_eikon_df}{fecha inicial de los datos a descargar}

\item{verbose}{argumento "verbose" pasado a Refinitiv::EikonGetData}

\item{raw_output}{argumento "raw_output" pasado a Refinitiv::EikonGetData}
}
\description{
Descargar datos de acciones vía Refinitiv Eikon, desde fecha especificada hasta hoy.
}
\examples{
descargar_datos_acciones_eikon(rics = c("SABE.MC", "BBVA.MC"), start_date_eikon_df=as.Date("2024-04-29"))
descargar_datos_acciones_eikon()
}
\keyword{acciones}
\keyword{datos}
\keyword{descargar}
\keyword{eikon}
