% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ejecutar_r.R
\name{ejecutar_r}
\alias{ejecutar_r}
\title{Ejecutar un archivo de R, comprobando resultado y notificando en Telegram}
\usage{
ejecutar_r(.bot_telegram, .archivo)
}
\arguments{
\item{plot_ggplot}{objeto ggplot con el gráfico original}
}
\description{
Ejecutar un archivo de R, comprobando resultado y notificando en Telegram
}
\examples{
ejecutar_r()

}
\keyword{generar}
\keyword{grafico}
\keyword{plotly}
