% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_percentage.R
\name{format_percentage}
\alias{format_percentage}
\title{Formatear números entre 0 y 1 como cadenas de texto que representan porcentajes.}
\usage{
format_percentage(x)
}
\arguments{
\item{x}{número a formatear}
}
\description{
Formatear números entre 0 y 1 como cadenas de texto que representan porcentajes.
}
\keyword{formatear}
\keyword{porcentaje}
