% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_grafico_basico.R
\name{generar_grafico_basico}
\alias{generar_grafico_basico}
\title{Generar un gráfico básico de líneas a partir de un tibble en formato long.}
\usage{
generar_grafico_basico(
  datos,
  .accuracy = 0.1,
  .scale = 1,
  .suffix = "",
  .big.mark = ".",
  .decimal.mark = ","
)
}
\arguments{
\item{datos}{tibble con los datos}

\item{.accuracy}{decimales, por defecto 0.1}

\item{.scale}{escala, por defecto 1}

\item{.suffix}{sufijo a añadir a los ejes y etiquetas}

\item{.big.mark}{marcador de miles, por defecto "."}

\item{.decimal.mark}{marcador de decimales, por defecto ","}
}
\description{
Generar un gráfico básico de líneas a partir de un tibble en formato long.
}
\examples{
generar_grafico_basico()
}
\keyword{generar}
\keyword{grafico}
\keyword{simple}
