% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_plotly.R
\name{generar_plotly}
\alias{generar_plotly}
\title{Generar un gráfico plotly a partir de un objeto ggplot, y asignarle un nombre en el espacio de nombres tal que: NOMBRE_GRAFICO_plotly_plt}
\usage{
generar_plotly(
  plot_ggplot,
  legendx = 0,
  legendy = -0.2,
  hoververtical = TRUE,
  label = "label",
  showlegend = TRUE,
  yaxisposition = "right",
  nombre_plotly_final = "",
  hoverhorizontal = FALSE,
  returnplot = FALSE
)
}
\arguments{
\item{plot_ggplot}{objeto ggplot con el gráfico original}

\item{legendx}{posición de la leyenda en el eje de abscisas}

\item{legendy}{posición de la leyenda en el eje de ordenadas}

\item{hoververtical}{hover sobre el gráfico en formato de línea vertical}

\item{label}{variable que contiene el label}

\item{showlegend}{mostrar o no la leyenda. Por defecto es TRUE}

\item{yaxisposition}{posición del eje de ordenadas. Por defecto es "right"}

\item{nombre_plotly_final}{nombre de la variable que contendrá el objeto plotly. Por defecto tiene formato NOMBRE_GRAFICO_plotly_plt}

\item{hoverhorizontal}{hover sobre el gráfico en formato de línea horizontal}
}
\description{
Generar un gráfico plotly a partir de un objeto ggplot, y asignarle un nombre en el espacio de nombres tal que: NOMBRE_GRAFICO_plotly_plt
}
\examples{
generar_plotly()
}
\keyword{generar}
\keyword{grafico}
\keyword{plotly}
