% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_png.R
\name{generar_png}
\alias{generar_png}
\title{A partir de un gr치fico ggplot, generar un archivo temporal con formato png.}
\usage{
generar_png(.grafico_a_png, width_png = 500, height_png = 500, res_png = 100)
}
\arguments{
\item{.grafico_a_png}{objeto ggplot con el gr치fico original}

\item{width_png}{anchura del png}

\item{height_png}{altura del png}

\item{res_png}{resoluci칩n del png}
}
\description{
A partir de un gr치fico ggplot, generar un archivo temporal con formato png.
}
\examples{
generar_png()
}
\keyword{archivo}
\keyword{generar}
\keyword{ggplot}
\keyword{grafico}
\keyword{png}
\keyword{temporal}
