% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniciar_telegram.R
\name{iniciar_telegram}
\alias{iniciar_telegram}
\title{Inicia el bot de telegram}
\usage{
iniciar_telegram()
}
\arguments{
\item{plot_ggplot}{objeto ggplot con el gráfico original}
}
\description{
Inicia el bot de telegram
}
\examples{
mensaje_telegram()

}
\keyword{generar}
\keyword{grafico}
\keyword{plotly}
