% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mensaje_telegram.R
\name{mensaje_telegram}
\alias{mensaje_telegram}
\title{Enviar un mensaje de Telegram vía bot de sganalisisfinanciero}
\usage{
mensaje_telegram(bot, mensaje)
}
\arguments{
\item{plot_ggplot}{objeto ggplot con el gráfico original}
}
\description{
Enviar un mensaje de Telegram vía bot de sganalisisfinanciero
}
\examples{
mensaje_telegram()

}
\keyword{generar}
\keyword{grafico}
\keyword{plotly}
